feat(events): rename TokenRefreshedEvent to UserObtainedTokenEvent, use this event when logging in and when refreshing
